refactor: change port type from u32 to u16 across codebase
TCP ports are inherently u16 (0-65535). Change ServerAddress::port(),
GrpcConfiguration.port, ClientProps.grpc_port, DEFAULT_SERVER_PORT,
and get_value_u32→get_value_u16 to use u16 consistently. Also updates
PollingServerAddress and NacosGrpcClient builder accordingly.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ src/api/constants.rs
@@ -1,6 +1,6 @@
 pub const DEFAULT_SERVER_ADDR: &str = "127.0.0.1:8848";
 
-pub const DEFAULT_SERVER_PORT: u32 = 8848;
+pub const DEFAULT_SERVER_PORT: u16 = 8848;
 
 /// Attention! "public" is "", it is recommended to customize the namespace with clear meaning.
 pub(crate) const DEFAULT_NAMESPACE: &str = "public";
